requirements.txt
ultralytics 
opencv-python
supervision
plotly
python-dotenv
openai
elevenlabs 
moviepy 
pydub
imageio-ffmpeg
tavily-python

